handle multiple move stop conditions

+++ motor-control/tests/test_limit_switch.cpp
@@ -37,7 +37,7 @@ TEST_CASE("Move with stop condition == limit switch") {
                      .acceleration = 0,
                      .group_id = 1,
                      .seq_id = 0,
-                     .stop_condition = MoveStopCondition::limit_switch};
+                     .stop_condition = static_cast<uint8_t>(MoveStopCondition::limit_switch)};
 
     GIVEN("the move is in progress") {
         test_objs.queue.try_write_isr(msg1);
@@ -78,7 +78,7 @@ TEST_CASE("Move with stop condition == limit switch, case 2") {
                      .acceleration = 0,
                      .group_id = 1,
                      .seq_id = 0,
-                     .stop_condition = MoveStopCondition::limit_switch};
+                     .stop_condition = static_cast<uint8_t>(MoveStopCondition::limit_switch)};
     GIVEN("the limit switch has not been triggered") {
         test_objs.queue.try_write_isr(msg1);
         test_objs.hw.set_mock_lim_sw(false);
@@ -114,7 +114,7 @@ TEST_CASE("Move with stop condition != limit switch") {
                      .acceleration = 0,
                      .group_id = 1,
                      .seq_id = 0,
-                     .stop_condition = MoveStopCondition::none};
+                     .stop_condition = static_cast<uint8_t>(MoveStopCondition::none)};
     GIVEN("Move with stop condition = none in progress") {
         test_objs.queue.try_write_isr(msg1);
         test_objs.hw.set_mock_lim_sw(true);
